[Debug] Change to use BigDecimal

+++ RaymentBackend/src/main/java/com/ma/raymond/rayment/controllers/DDController.java
@@ -8,6 +8,8 @@ import com.ma.raymond.rayment.services.DDService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
+
 @RestController
 @RequestMapping("dd")
 public class DDController {
@@ -15,7 +17,7 @@ public class DDController {
     DDService ddService;
 
     @PutMapping("/dd")
-    public ResponseDTO ct(@RequestParam("from_acc_id") Integer fromAccId, @RequestParam("to_acc_id") Integer toAccId, @RequestParam("curr") String curr, @RequestParam("amount") double amount) throws AccountNotFoundException, InsufficientFundException {
+    public ResponseDTO ct(@RequestParam("from_acc_id") Integer fromAccId, @RequestParam("to_acc_id") Integer toAccId, @RequestParam("curr") String curr, @RequestParam("amount") BigDecimal amount) throws AccountNotFoundException, InsufficientFundException {
         ddService.dd(fromAccId,toAccId,curr,amount);
         return new ResponseDTO("", String.format("Direct Debit %s %.2f OK ",curr,amount), "", "temp instance", "helpUrl", null);
     }
